Dockerfile: use debian image again
As we depend on conntrack, we need to ensure that we have all the
libraries related to conntrack.

The reason for having all commands in the same "RUN" instructions is to
ensure that the layer is as small as possible.

Now we have a working bs image, which is 150MB (it's not as good as the
cirros one, where we got 18MB, but it's better than the previous one
based on the golang image, which had 800MB in total).

I will probably give it another try in the future. Maybe Docker will
give us more details on network stats and we will be able to drop
conntrack? Or can we statically build conntrack? :)