status: fix test for Go 1.6
Looks like StatusCode is no longer 200 on Response. This code was just
wrong semantically, but I'm not sure if Go should really change that.
Well, I don't care enough for raising an issue, as I don't wanna keep
this bad behavior on the Go standard library.

Should fix the build.