main,metric,status: allow bs to run without tsuru endpoint
This makes bs able to run without having a tsuru endpoint api. In
this kind of configuration status won't be reported (since there is
no tsuru api to report to) and it's not possible to have tsuru as
one of the log backends.

The metric runner/reporter are now to responsible for keeping bs running
and bs will exit if it fails to setup the metric runner.